Добавил зкрытие модалки по esc
