+++ src/components/modal/modal-overlay/modal-overlay.module.css
@@ -1,0 +1,18 @@
+.overlay{
+    position: fixed;
+    top: 0;
+    bottom: 0;
+    right: 0;
+    left: 0;
+    overflow: hidden;
+    overflow-y: auto;
+    background: rgba(0,0,0,0.60);
+    -webkit-overflow-scrolling: touch;
+    display: flex;
+    flex-flow: column nowrap;
+    justify-content: center;
+    align-items: center;
+    z-index: 99;
+}
+
+
