+++ src/components/modal/modal.module.css
@@ -1,19 +1,3 @@
-.overlay{
-    position: fixed;
-    top: 0;
-    bottom: 0;
-    right: 0;
-    left: 0;
-    overflow: hidden;
-    overflow-y: auto;
-    background: rgba(0,0,0,0.60);
-    -webkit-overflow-scrolling: touch;
-    display: flex;
-    flex-flow: column nowrap;
-    justify-content: center;
-    align-items: center;
-    z-index: 99;
-}
 
 .modal{
 
@@ -31,6 +15,7 @@
     box-shadow: 0 24px 32px rgba(0, 0, 0, 0.04), 0 16px 24px rgba(0, 0, 0, 0.04), 0 4px 8px rgba(0, 0, 0, 0.04), 0 0 1px rgba(0, 0, 0, 0.04);
     border-radius: 40px;
 }
+
 .closeCross{
     position: absolute;
     top: 60px;
@@ -47,4 +32,3 @@
     width: 31px;
     height: 31px;
 }
-
